ml_model/requirements.txt
flask==2.3.3
timm==0.9.5
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
requests==2.31.0
